Completed authorisation and authenthication of user closes #8
I have completed authorisation of users where i used a jwt token allow
those who have the jwt token to access the profile. I used
authentication for the login and after login there will be a jwt token
created. The jwt token will the be verified for specific routes to enter
and it will be saved in the request path for further usage in the routed
page